Use longs for most pojo. Replace string literals in jersey annotations with predefined properties.

+++ src/main/java/com/clevertap/model/GetCursorRequest.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Getter
 public class GetCursorRequest {
 
-  private final int batchSize;
+  private final long batchSize;
   private final boolean requireAppFields;
   private final boolean requireEventsSummary;
   private final boolean requireCustomProfileProperties;
